+++ fake-cucumber/javascript/src/Cucumber.ts
@@ -3,7 +3,7 @@ import { MessageNotifier } from './types'
 import TestPlan from './TestPlan'
 import IStepDefinition from './IStepDefinition'
 import IHook from './IHook'
-import { GherkinQuery } from '@cucumber/gherkin'
+import { Query } from '@cucumber/gherkin'
 import { ParameterType } from '@cucumber/cucumber-expressions'
 import IClock from './IClock'
 import { MakeErrorMessage } from './ErrorMessageGenerator'
@@ -19,7 +19,7 @@ export default class Cucumber {
     private readonly undefinedParameterTypes: messages.IEnvelope[],
     private readonly beforeHooks: IHook[],
     private readonly afterHooks: IHook[],
-    private readonly gherkinQuery: GherkinQuery,
+    private readonly gherkinQuery: Query,
     private readonly newId: IdGenerator.NewId,
     private readonly clock: IClock,
     private readonly makeErrorMessage: MakeErrorMessage