+++ query/javascript/src/Query.ts
@@ -1,28 +1,25 @@
-import { StrictArrayMultimap, StrictMap } from '@cucumber/gherkin'
 import { messages, TimeConversion } from '@cucumber/messages'
+import { ArrayMultimap } from '@teppeis/multimaps'
 
 export default class Query {
-  private readonly testStepResultByPickleId = new StrictArrayMultimap<
+  private readonly testStepResultByPickleId = new ArrayMultimap<
     string,
     messages.ITestStepResult
   >()
-  private readonly testStepResultsByPickleStepId = new StrictArrayMultimap<
+  private readonly testStepResultsByPickleStepId = new ArrayMultimap<
     string,
     messages.ITestStepResult
   >()
-  private readonly testStepById = new StrictMap<
-    string,
-    messages.TestCase.ITestStep
-  >()
-  private readonly pickleIdByTestStepId = new StrictMap<string, string>()
-  private readonly pickleStepIdByTestStepId = new StrictMap<string, string>()
+  private readonly testStepById = new Map<string, messages.TestCase.ITestStep>()
+  private readonly pickleIdByTestStepId = new Map<string, string>()
+  private readonly pickleStepIdByTestStepId = new Map<string, string>()
 
-  private readonly attachmentsByPickleStepId = new StrictArrayMultimap<
+  private readonly attachmentsByPickleStepId = new ArrayMultimap<
     string,
     messages.IAttachment
   >()
 
-  private readonly stepMatchArgumentsListsByPickleStepId = new StrictMap<
+  private readonly stepMatchArgumentsListsByPickleStepId = new Map<
     string,
     messages.TestCase.TestStep.IStepMatchArgumentsList[]
   >()
@@ -76,7 +73,7 @@ export default class Query {
     if (pickleStepIds.length === 0) {
       return [
         new messages.TestStepResult({
-          status: messages.TestStepResult.Status.SKIPPED,
+          status: messages.TestStepResult.Status.UNKNOWN,
           duration: TimeConversion.millisecondsToDuration(0),
         }),
       ]
@@ -101,7 +98,7 @@ export default class Query {
     if (pickleIds.length === 0) {
       return [
         new messages.TestStepResult({
-          status: messages.TestStepResult.Status.UNDEFINED,
+          status: messages.TestStepResult.Status.UNKNOWN,
           duration: TimeConversion.millisecondsToDuration(0),
         }),
       ]
@@ -126,7 +123,7 @@ export default class Query {
     return (
       testStepResults.sort((r1, r2) => r2.status - r1.status)[0] ||
       new messages.TestStepResult({
-        status: messages.TestStepResult.Status.SKIPPED,
+        status: messages.TestStepResult.Status.UNKNOWN,
         duration: TimeConversion.millisecondsToDuration(0),
       })
     )
@@ -155,7 +152,7 @@ export default class Query {
    */
   public getStepMatchArgumentsLists(
     pickleStepId: string
-  ): messages.TestCase.TestStep.IStepMatchArgumentsList[] {
+  ): messages.TestCase.TestStep.IStepMatchArgumentsList[] | undefined {
     return this.stepMatchArgumentsListsByPickleStepId.get(pickleStepId)
   }
 }